Update frontend-checklist.md
`Pagination with `rel=”next”` and `rel=”prev”` attributes.`

From what I can tell, this isn't needed. 

To be clear (because it wasn't clear), this seems to be a `<head>` element, not something you put on your in-page pagination links. i.e.:
```
<head>
  <link rel="prev" href="some-other-webpage.html">
```

It used to be used as an SEO thing, but a) google stopped supporting this YEARS ago and b) we don't tend to build pages like this.